Microsoft Azure Cloud Native - Lab 1/requirements.txt
streamlit
azure-storage-blob
pymssql
dotenv
